+++ mitra_api/src/mastodon_api/notifications/types.rs
@@ -21,9 +21,10 @@ use crate::mastodon_api::{
 
 fn default_page_size() -> PageSize { PageSize::new(20) }
 
-/// https://docs.joinmastodon.org/methods/notifications/
+// https://docs.joinmastodon.org/methods/notifications/
 #[derive(Deserialize)]
 pub struct NotificationQueryParams {
+    pub min_id: Option<i32>,
     pub max_id: Option<i32>,
 
     #[serde(default = "default_page_size")]
@@ -36,7 +37,7 @@ pub struct EmojiReaction {
     emoji: Option<CustomEmoji>,
 }
 
-/// https://docs.joinmastodon.org/entities/notification/
+// https://docs.joinmastodon.org/entities/notification/
 #[derive(Serialize)]
 pub struct Notification {
     pub id: String,